Pianoroll (Player Inputs)
This version syncs with savestates.
